# contracts/openapi/bff-gateway.openapi.yaml
openapi: 3.0.3
info:
  title: SwasthKart BFF Gateway
  description: |
    Frontend-facing Backend-For-Frontend API gateway.
    Handles authentication via HttpOnly JWT cookies, aggregates data from
    internal microservices, and serves enriched responses for the web/mobile clients.

    Admin endpoints under /api/v1/admin/* are proxied to the admin-service — see admin-service.openapi.yaml.
  version: 1.0.0
  contact:
    name: SwasthKart Engineering
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/ui/profile:
    get:
      tags: [Profile]
      summary: Get current user's profile
      operationId: getProfile
      security:
        - CookieAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"

    put:
      tags: [Profile]
      summary: Update current user's profile
      operationId: updateProfile
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                  example: "Aarav Sharma"
      responses:
        "200":
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

  # ─── UI Authenticated — Addresses ──────────────────────────────────────
  /api/v1/ui/addresses:
    get:
      tags: [Addresses]
      summary: List user's addresses
      operationId: listAddresses
      security:
        - CookieAuth: []
      responses:
        "200":
          description: List of addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [Addresses]
      summary: Add a new address
      operationId: createAddress
      security:
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "201":
          description: Address created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
  /api/v1/ui/addresses/{id}:
    put:
      tags: [Addresses]
      summary: Update an address
      operationId: updateAddress
      security:
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Address ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "200":
          description: Updated address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"

    delete:
      tags: [Addresses]
      summary: Delete an address
      operationId: deleteAddress
      security:
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Address ID
      responses:
        "204":
          description: Address deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/ui/addresses/{id}/default:
    put:
      tags: [Addresses]
      summary: Set address as default
      operationId: setDefaultAddress
      security:
        - CookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Address ID
      responses:
        "200":
          description: Address set as default
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─── Admin (proxied) ──────────────────────────────────────────────────
  /api/v1/admin/dashboard:
    get:
      tags: [Admin]
      summary: "[Proxied] Get admin dashboard metrics"
      description: Proxied to admin-service. See admin-service.openapi.yaml for full details.
      operationId: adminGetDashboard
      security:
        - CookieAuth: []
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/inventory:
    get:
      tags: [Admin]
      summary: "[Proxied] List inventory"
      description: Proxied to admin-service. See admin-service.openapi.yaml for full details.
      operationId: adminListInventory
      security:
        - CookieAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Paginated inventory items
          content:
            application/json:
              schema:
                type: object
                required: [data, pagination]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/InventoryItem"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/inventory/{sku}:
    put:
      tags: [Admin]
      summary: "[Proxied] Update inventory"
      description: Proxied to admin-service. See admin-service.openapi.yaml for full details.
      operationId: adminUpdateInventory
      security:
        - CookieAuth: []
      parameters:
        - name: sku
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [total_qty]
              properties:
                total_qty:
                  type: integer
                  minimum: 0
      responses:
        "200":
          description: Updated inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: swasthkart_token
      description: HttpOnly JWT cookie set by POST /api/v1/auth/otp/verify
  parameters:
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-indexed)
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: UNAUTHORIZED
              message: Authentication required
              details: {}
    Forbidden:
      description: Insufficient permissions — admin role required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: FORBIDDEN
              message: Admin role required
              details: {}
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
              details: {}
    UnprocessableEntity:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: VALIDATION_ERROR
              message: Invalid input
              details: {}
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid input provided
            details:
              type: object
              additionalProperties: true
    Pagination:
      type: object
      required: [page, page_size, total_count, total_pages]
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        page_size:
          type: integer
          minimum: 1
          example: 20
        total_count:
          type: integer
          minimum: 0
          example: 150
        total_pages:
          type: integer
          minimum: 0
          example: 8

    # ─── Catalog ───────────────────────────────────────────────────────
...
    User:
      type: object
      required: [id, phone, name, role, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        phone:
          type: string
          example: "+919876543210"
        name:
          type: string
          example: "Aarav Sharma"
        role:
          type: string
          enum: [USER, ADMIN]
          example: USER
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Address:
      type: object
      required: [id, user_id, label, line1, city, state, pincode, is_default, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        label:
          type: string
          example: Home
        line1:
          type: string
          example: "42, Lotus Apartments"
        line2:
          type: string
          nullable: true
          example: "MG Road, Koramangala"
        city:
          type: string
          example: Bengaluru
        state:
          type: string
          example: Karnataka
        pincode:
          type: string
          example: "560034"
        is_default:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AddressInput:
      type: object
      required: [label, line1, city, state, pincode]
      properties:
        label:
          type: string
          minLength: 1
          maxLength: 50
          example: Home
        line1:
          type: string
          minLength: 1
          maxLength: 200
          example: "42, Lotus Apartments"
        line2:
          type: string
          nullable: true
          maxLength: 200
          example: "MG Road, Koramangala"
        city:
          type: string
          minLength: 1
          maxLength: 100
          example: Bengaluru
        state:
          type: string
          minLength: 1
          maxLength: 100
          example: Karnataka
        pincode:
          type: string
          pattern: '^\d{6}$'
          example: "560034"

    # ─── Admin (proxied schemas) ─────────────────────────────────────
    DashboardResponse:
      type: object
      required: [total_orders, orders_today, revenue_today_paise, pending_deliveries]
      properties:
        total_orders:
          type: integer
          example: 12450
        orders_today:
          type: integer
          example: 37
        revenue_today_paise:
          type: integer
          description: Revenue today in paise
          example: 4523500
        pending_deliveries:
          type: integer
          example: 15
    InventoryItem:
      type: object
      required: [sku, product_id, total_qty, available_qty, reserved_qty]
      properties:
        sku:
          type: string
          example: "RICE-BAS-1KG"
        product_id:
          type: string
          format: uuid
        total_qty:
          type: integer
          minimum: 0
          example: 500
        available_qty:
          type: integer
          minimum: 0
          example: 420
        reserved_qty:
          type: integer
          minimum: 0
          example: 80